Fix the issue in CreateConstantOnes which cause the float 16 model test failure. The CreateConstantOnes doesn't create the data with correct data type for float 16. For float 16, should use cuda::ToCudaType<MLFloat16>::FromFloat(1.0f) to create the data.
