+++ onnxruntime/test/onnx/main.cc
@@ -351,7 +351,6 @@ int real_main(int argc, char* argv[]) {
 
   broken_tests["fp16_tiny_yolov2"] = "unknown failure on CUDA";
   broken_tests["fp16_shufflenet"] = "unknown failure on CUDA";
-  broken_tests["fp16_inception_v1"] = "unknown failure on CUDA";
 #endif
 
   int result = 0;
